requirements.txt
cairocffi==0.8.1
CairoSVG==2.1.3
cffi==1.11.5
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.0.2
django-admin-interface==0.6.3
django-ajax-selects==1.7.0
django-colorfield==0.1.14
django-crispy-forms==1.7.0
django-debug-toolbar==1.9.1
django-filter==2.0.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-money==0.12.3
django-multiselectfield==0.1.8
django-qr-code==1.0.0
django-select2-rocks==0.7.0
django-simple-history==2.2.0
django-widget-tweaks==1.4.2
django-wkhtmltopdf==3.2.0
djangorestframework==3.8.2
ez-setup==0.9
html5lib==1.0.1
httplib2==0.11.3
Markdown==2.6.11
nose==1.3.7
pdfkit==0.6.1
pdfrw==0.4
Pillow==5.2.0
py-moneyed==0.7.0
pycparser==2.18
PyPDF2==1.26.0
Pyphen==0.9.4
python-decouple==3.1
pytz==2018.3
qrcode==6.0
reportlab==3.5.1
six==1.11.0
sqlparse==0.2.4
sweetify==1.0.3
tinycss2==0.6.1
Unipath==1.1
WeasyPrint==0.42.3
webencodings==0.5.1
xhtml2pdf==0.2.2
